requirements.txt
Babel~=2.14.0
PyLaTeX~=1.4.2
python_dateutil~=2.9.0.post0
PyYAML~=6.0.1
tqdm~=4.66.4
pytest~=8.3.3
python-dateutil~=2.9.0.post0
